Patially update code
