+++ src/index.ts
@@ -1,14 +1,17 @@
 import * as cluster from 'cluster'
-import { Worker } from './modules/worker/worker'
-import { BrokerServer } from './modules/broker/broker'
-import { CustomObject, Configurations, Options, Message, logReady, logError, logWarning, generateKey } from './utils/utils'
+import { logReady, logError, logWarning, generateKey } from './utils/functions'
+import { Configurations, Options, CustomObject, Message } from './utils/interfaces'
+
+declare const process: any
 
 export default class ClusterWS {
+    private options: Options
+
     constructor(configurations: Configurations) {
         if ({}.toString.call(configurations.worker) !== '[object Function]')
             return logError('Worker must be provided and it must be a function \n')
 
-        const options: Options = {
+        this.options = {
             port: configurations.port || (configurations.tlsOptions ? 443 : 80),
             worker: configurations.worker,
             workers: configurations.workers || 1,
@@ -22,73 +25,78 @@ export default class ClusterWS {
         }
 
         if (!configurations.brokersPorts)
-            for (let i: number = 0; i < options.brokers; i++) options.brokersPorts.push(9400 + i)
-
-        if (options.brokersPorts.length < options.brokers)
+            for (let i: number = 0; i < this.options.brokers; i++) this.options.brokersPorts.push(9400 + i)
+        if (this.options.brokersPorts.length < this.options.brokers)
             return logError('Number of broker ports is less than number of brokers \n')
 
-        cluster.isMaster ? MasterProcess(options) : WorkerProcess(options)
+        cluster.isMaster ? this.masterProcess() : this.workerProcess()
     }
-}
 
-function MasterProcess(options: Options): void {
-    let loading: boolean = true
+    private masterProcess(): void {
+        let loaded: boolean = false
+        const key: string = generateKey(16)
+        const brokersReady: CustomObject = {}
+        const workersReady: CustomObject = {}
 
-    const key: string = generateKey(16)
-    const brokersReady: CustomObject = {}
-    const workersReady: CustomObject = {}
+        if (this.options.horizontalScaleOptions && this.options.horizontalScaleOptions.masterOptions)
+            launchProcess('Scaler', -1)
+        else for (let i: number = 0; i < this.options.brokers; i++)
+            launchProcess('Broker', i)
 
-    if (options.horizontalScaleOptions && options.horizontalScaleOptions.masterPort)
-        launchProcess('Scaler', -1)
-    else for (let i: number = 0; i < options.brokers; i++) launchProcess('Broker', i)
+        function launchProcess(processName: string, processId: number): void {
+            let newProcess: cluster.Worker = cluster.fork()
 
-    function launchProcess(processName: string, processID: number): void {
-        let newProcess: cluster.Worker = cluster.fork()
-        newProcess.on('message', (message: Message) => message.event === 'READY' && ready(processName, processID, message.pid))
-        newProcess.on('exit', (): void => {
-            logError(processName + ' has been disconnected \n')
-            if (options.restartWorkerOnFail) {
-                logWarning(processName + ' is restarting \n')
-                launchProcess(processName, processID)
+            newProcess.on('message', (message: Message) =>
+                message.event === 'READY' && ready(processName, processId, message.pid))
+            newProcess.on('exit', () => {
+                logError(processName + ' is closed \n')
+                if (this.options.restartWorkerOnFail) {
+                    logWarning(processName + ' is restarting \n')
+                    launchProcess(processName, processId)
+                }
+                newProcess = null
+            })
+            newProcess.send({ key, processId, processName })
+        }
+
+        function ready(processName: string, processId: number, pid: number): void {
+            if (loaded)
+                return logReady(processName + ' PID ' + pid + ' has restarted')
+
+            if (processName === 'Worker')
+                workersReady[processId] = '\tWorker: ' + processId + ', PID ' + pid
+
+            if (processName === 'Scaler')
+                for (let i: number = 0; i < this.options.brokers; i++)
+                    launchProcess('Broker', i)
+
+            if (processName === 'Broker') {
+                brokersReady[processId] = '>>>  Broker on: ' + this.options.brokersPorts[processId] + ', PID ' + pid
+                if (Object.keys(brokersReady).length === this.options.brokers)
+                    for (let i: number = 0; i < this.options.workers; i++)
+                        launchProcess('Worker', i)
             }
-            newProcess = null
+
+            if (Object.keys(brokersReady).length === this.options.brokers && Object.keys(workersReady).length === this.options.workers) {
+                loaded = true
+                logReady('>>>  Master on: ' + this.options.port + ', PID: ' + process.pid + (this.options.tlsOptions ? ' (secure)' : ''))
+                for (const indexKey in brokersReady) brokersReady[indexKey] && logReady(brokersReady[indexKey])
+                for (const indexKey in workersReady) workersReady[indexKey] && logReady(workersReady[indexKey])
+            }
+        }
+    }
+
+    private workerProcess(): void {
+        process.on('message', (message: Message): void => {
+            switch (message.processName) {
+                case 'Broker':
+                case 'Worker':
+                case 'Scaler':
+            }
         })
-        newProcess.send({ key, processID, processName })
+        process.on('uncaughtException', (err: Error): void => {
+            logError('PID: ' + process.pid + '\n' + err.stack + '\n')
+            return process.exit()
+        })
     }
-
-    function ready(processName: string, processID: number, pid: number): void {
-        if (!loading) return logReady(processName + ' PID ' + pid + ' has restarted')
-        if (processName === 'Worker') workersReady[processID] = '       Worker: ' + processID + ', PID ' + pid
-        if (processName === 'Scaler')
-            for (let i: number = 0; i < options.brokers; i++) launchProcess('Broker', i)
-        if (processName === 'Broker') {
-            brokersReady[processID] = '>>> Broker on: ' + options.brokersPorts[processID] + ', PID ' + pid
-            if (Object.keys(brokersReady).length === options.brokers)
-                for (let i: number = 0; i < options.workers; i++) launchProcess('Worker', i)
-        }
-        if (Object.keys(brokersReady).length === options.brokers && Object.keys(workersReady).length === options.workers) {
-            loading = false
-            logReady('>>> Master on: ' + options.port + ', PID: ' + process.pid + (options.tlsOptions ? ' (secure)' : ''))
-            for (const indexKey in brokersReady) brokersReady[indexKey] && logReady(brokersReady[indexKey])
-            for (const indexKey in workersReady) workersReady[indexKey] && logReady(workersReady[indexKey])
-        }
-    }
-}
-
-function WorkerProcess(options: Options): void {
-    process.on('message', (message: Message): void | Worker => {
-        switch (message.processName) {
-            case 'Broker': return BrokerServer(options.brokersPorts[message.processID], message.key, options.horizontalScaleOptions)
-            case 'Worker': return new Worker(options, message.key)
-            case 'Scaler': return options.horizontalScaleOptions &&
-                BrokerServer(options.horizontalScaleOptions.masterPort,
-                    options.horizontalScaleOptions.key || '',
-                    options.horizontalScaleOptions,
-                    'Scaler')
-        }
-    })
-    process.on('uncaughtException', (err: Error): void => {
-        logError('PID: ' + process.pid + '\n' + err.stack + '\n')
-        return process.exit()
-    })
 }
