+++ src/modules/broker/client.ts
@@ -1,0 +1,5 @@
+import * as WebSocket from 'uws'
+
+export function BrokerClient(): void {
+
+}
