+++ src/utils/interfaces.ts
@@ -1,21 +1,3 @@
-import * as crypto from 'crypto'
-
-export function logError<T>(data: T): any {
-    return console.log('\x1b[31m%s\x1b[0m', data)
-}
-
-export function logReady<T>(data: T): any {
-    return console.log('\x1b[36m%s\x1b[0m', data)
-}
-
-export function logWarning<T>(data: T): any {
-    return console.log('\x1b[33m%s\x1b[0m', data)
-}
-
-export function generateKey(length: number): string {
-    return crypto.randomBytes(Math.ceil(length / 2)).toString('hex').slice(0, length)
-}
-
 export type Message = any
 export type Listener = (...args: any[]) => void
 export type WorkerFunction = () => void
@@ -33,9 +15,11 @@ export interface TlsOptions {
 }
 
 export interface HorizontalScaleOptions {
-    masterPort?: number
-    masterTlsOptions?: TlsOptions
-    mastersUrls?: string[]
+    masterOptions?: {
+        port: number
+        tlsOptions?: TlsOptions
+    }
+    brokersUrls?: string[]
     key?: string
 }
