+++ src1/index.ts
@@ -1,84 +1,94 @@
 import * as cluster from 'cluster'
-
 import { Worker } from './modules/worker/worker'
-import { Broker } from './modules/broker/broker'
-import { CustomObject, Options, Configurations, logError, logWarning, logReady, randomString } from './utils/utils'
-
-declare const process: any
+import { BrokerServer } from './modules/broker/broker'
+import { CustomObject, Configurations, Options, Message, logReady, logError, logWarning, generateKey } from './utils/utils'
 
 export default class ClusterWS {
     constructor(configurations: Configurations) {
         if ({}.toString.call(configurations.worker) !== '[object Function]')
-            return logError('Worker must be provided and it must be a function \n \n')
+            return logError('Worker must be provided and it must be a function \n')
 
         const options: Options = {
             port: configurations.port || (configurations.tlsOptions ? 443 : 80),
             worker: configurations.worker,
             workers: configurations.workers || 1,
+            brokers: configurations.brokers || 1,
             useBinary: configurations.useBinary || false,
-            brokerPort: configurations.brokerPort || 9346,
+            brokersPorts: configurations.brokersPorts || [],
             tlsOptions: configurations.tlsOptions || false,
-            scaleOptions: configurations.scaleOptions || false,
             pingInterval: configurations.pingInterval || 20000,
-            restartWorkerOnFail: configurations.restartWorkerOnFail || false
+            restartWorkerOnFail: configurations.restartWorkerOnFail || false,
+            horizontalScaleOptions: configurations.horizontalScaleOptions || false
         }
-        cluster.isMaster ? ClusterWS.master(options) : ClusterWS.worker(options)
+
+        if (!configurations.brokersPorts)
+            for (let i: number = 0; i < options.brokers; i++) options.brokersPorts.push(9400 + i)
+
+        if (options.brokersPorts.length < options.brokers)
+            return logError('Number of broker ports is less than number of brokers \n')
+
+        cluster.isMaster ? MasterProcess(options) : WorkerProcess(options)
     }
+}
 
-    private static master(options: Options): void {
-        let isReady: boolean = false
+function MasterProcess(options: Options): void {
+    let loading: boolean = true
 
-        const key: string = randomString(16)
-        const processes: CustomObject = {}
+    const key: string = generateKey(16)
+    const brokersReady: CustomObject = {}
+    const workersReady: CustomObject = {}
 
-        options.scaleOptions && options.scaleOptions.master ?
-            launchProcess('Scaler', -1) : launchProcess('Broker', 0)
+    if (options.horizontalScaleOptions && options.horizontalScaleOptions.masterPort)
+        launchProcess('Scaler', -1)
+    else for (let i: number = 0; i < options.brokers; i++) launchProcess('Broker', i)
 
-        function launchProcess(processName: string, processId: number): void {
-            const runningProcess: cluster.Worker = cluster.fork()
-            runningProcess.send({ processName, key })
-            runningProcess.on('message', (message: CustomObject): void =>
-                message.event === 'READY' && ready(processName, processId, message.pid))
-            runningProcess.on('exit', (): void => {
-                logWarning(processName + ' has been disconnected \n')
-                if (options.restartWorkerOnFail) {
-                    logWarning(processName + ' is restarting \n')
-                    launchProcess(processName, processId)
-                }
-            })
-        }
-
-        function ready(processName: string, processId: number, pid: number): void {
-            if (isReady) return logReady(processName + ' has restarted')
-            if (processName === 'Scaler') return launchProcess('Broker', 0)
-            if (processName === 'Broker') for (let i: number = 1; i <= options.workers; i++) launchProcess('Worker', i)
-
-            processes[processId] = pid
-            if (Object.keys(processes).length === options.workers + 1) {
-                isReady = true
-                logReady('>>> Master on: ' + options.port + ', PID: ' + process.pid + (options.tlsOptions ? ' (secure)' : ''))
-                for (const index in processes)
-                    processes[index] && index === '0' ?
-                        logReady('>>> Broker on: ' + options.brokerPort + ', PID ' + processes[index]) :
-                        logReady('       Worker: ' + index + ', PID ' + processes[index])
+    function launchProcess(processName: string, processID: number): void {
+        let newProcess: cluster.Worker = cluster.fork()
+        newProcess.on('message', (message: Message) => message.event === 'READY' && ready(processName, processID, message.pid))
+        newProcess.on('exit', (): void => {
+            logError(processName + ' has been disconnected \n')
+            if (options.restartWorkerOnFail) {
+                logWarning(processName + ' is restarting \n')
+                launchProcess(processName, processID)
             }
-        }
+            newProcess = null
+        })
+        newProcess.send({ key, processID, processName })
     }
 
-    private static worker(options: Options): void {
-        process.on('message', (message: CustomObject): void | Worker => {
-            switch (message.processName) {
-                case 'Worker':
-                    return new Worker(options, message.key)
-                case 'Broker':
-                    return Broker.Server(options.brokerPort, message.key, options.scaleOptions)
-                case 'Scaler':
-                    return options.scaleOptions && Broker.Server(options.scaleOptions.port, options.scaleOptions.key || '')
-            }
-        })
-        process.on('uncaughtException', (err: Error): void => {
-            logError('PID: ' + process.pid + '\n' + err.stack + '\n')
-            return process.exit()
-        })
+    function ready(processName: string, processID: number, pid: number): void {
+        if (!loading) return logReady(processName + ' PID ' + pid + ' has restarted')
+        if (processName === 'Worker') workersReady[processID] = '       Worker: ' + processID + ', PID ' + pid
+        if (processName === 'Scaler')
+            for (let i: number = 0; i < options.brokers; i++) launchProcess('Broker', i)
+        if (processName === 'Broker') {
+            brokersReady[processID] = '>>> Broker on: ' + options.brokersPorts[processID] + ', PID ' + pid
+            if (Object.keys(brokersReady).length === options.brokers)
+                for (let i: number = 0; i < options.workers; i++) launchProcess('Worker', i)
+        }
+        if (Object.keys(brokersReady).length === options.brokers && Object.keys(workersReady).length === options.workers) {
+            loading = false
+            logReady('>>> Master on: ' + options.port + ', PID: ' + process.pid + (options.tlsOptions ? ' (secure)' : ''))
+            for (const indexKey in brokersReady) brokersReady[indexKey] && logReady(brokersReady[indexKey])
+            for (const indexKey in workersReady) workersReady[indexKey] && logReady(workersReady[indexKey])
+        }
     }
+}
+
+function WorkerProcess(options: Options): void {
+    process.on('message', (message: Message): void | Worker => {
+        switch (message.processName) {
+            case 'Broker': return BrokerServer(options.brokersPorts[message.processID], message.key, options.horizontalScaleOptions)
+            case 'Worker': return new Worker(options, message.key)
+            case 'Scaler': return options.horizontalScaleOptions &&
+                BrokerServer(options.horizontalScaleOptions.masterPort,
+                    options.horizontalScaleOptions.key || '',
+                    options.horizontalScaleOptions,
+                    'Scaler')
+        }
+    })
+    process.on('uncaughtException', (err: Error): void => {
+        logError('PID: ' + process.pid + '\n' + err.stack + '\n')
+        return process.exit()
+    })
 }
