+++ src1/utils/utils.ts
@@ -1,5 +1,22 @@
 import * as crypto from 'crypto'
 
+export function logError<T>(data: T): any {
+    return console.log('\x1b[31m%s\x1b[0m', data)
+}
+
+export function logReady<T>(data: T): any {
+    return console.log('\x1b[36m%s\x1b[0m', data)
+}
+
+export function logWarning<T>(data: T): any {
+    return console.log('\x1b[33m%s\x1b[0m', data)
+}
+
+export function generateKey(length: number): string {
+    return crypto.randomBytes(Math.ceil(length / 2)).toString('hex').slice(0, length)
+}
+
+export type Message = any
 export type Listener = (...args: any[]) => void
 export type WorkerFunction = () => void
 
@@ -15,49 +32,35 @@ export interface TlsOptions {
     passphrase?: string
 }
 
-export interface ScaleOptions {
-    port: number
-    url?: string
+export interface HorizontalScaleOptions {
+    masterPort?: number
+    masterTlsOptions?: TlsOptions
+    mastersUrls?: string[]
     key?: string
-    master?: boolean
 }
 
 export interface Configurations {
     worker: WorkerFunction
     port?: number
     workers?: number
+    brokers?: number
     useBinary?: boolean
-    brokerPort?: number
+    brokersPorts?: number[]
     tlsOptions?: TlsOptions
-    scaleOptions?: ScaleOptions
     pingInterval?: number
     restartWorkerOnFail?: boolean
+    horizontalScaleOptions?: HorizontalScaleOptions
 }
 
 export interface Options {
     worker: WorkerFunction
     port: number
     workers: number
+    brokers: number
     useBinary: boolean
-    brokerPort: number
+    brokersPorts: number[]
     tlsOptions: TlsOptions | false
-    scaleOptions: ScaleOptions | false
     pingInterval: number
     restartWorkerOnFail: boolean
-}
-
-export function logError<T>(data: T): any {
-    return console.log('\x1b[31m%s\x1b[0m', data)
-}
-
-export function logReady<T>(data: T): any {
-    return console.log('\x1b[36m%s\x1b[0m', data)
-}
-
-export function logWarning<T>(data: T): any {
-    return console.log('\x1b[33m%s\x1b[0m', data)
-}
-
-export function randomString(length: number): string {
-    return crypto.randomBytes(Math.ceil(length / 2)).toString('hex').slice(0, length)
+    horizontalScaleOptions: HorizontalScaleOptions | false
 }
